requirements.txt
# Core dependencies
streamlit==1.35.0
python-dotenv==1.0.0

# LangChain and AI
langchain==0.2.14
langchain-community==0.2.12
langchain-groq==0.1.9

# Data processing
pandas==2.0.2

# Vector database
chromadb==0.4.24

# AWS Cognito authentication
boto3==1.35.0

# Document processing (for WebBaseLoader)
unstructured==0.14.6

# Resume parsing dependencies
pypdf>=6.2.0
python-docx==1.1.0
